Remove signers
