--- gems/smithy-client/lib/smithy-client/signer.rb
@@ -1,16 +1,0 @@
-# frozen_string_literal: true
-
-module Smithy
-  module Client
-    # Base class for all Signer classes.
-    class Signer
-      def sign(_options = {})
-        raise NotImplementedError
-      end
-
-      def reset(_options = {})
-        raise NotImplementedError
-      end
-    end
-  end
-end
