--- gems/smithy-client/lib/smithy-client/signers/anonymous.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Smithy
-  module Client
-    module Signers
-      # A signer that does not sign requests.
-      class Anonymous < Signer
-        def sign(context); end
-        def reset(context); end
-      end
-    end
-  end
-end
