+++ gems/smithy/lib/smithy/welds/auth/anonymous_auth.rb
@@ -16,6 +16,16 @@ module Smithy
           Smithy::Client::Plugins::AnonymousAuth => { require_path: 'smithy-client/plugins/anonymous_auth' }
         }
       end
+
+      def add_auth_schemes
+        {
+          'smithy.api#noAuth' => {
+            auth_scheme_config_option: :anonymous_auth_scheme,
+            identity_provider_config_option: :anonymous_provider,
+            identity_type: Smithy::Client::Identities::Anonymous
+          }
+        }
+      end
     end
   end
 end
