+++ gems/smithy/lib/smithy/welds/auth/http_api_key_auth.rb
@@ -19,6 +19,16 @@ module Smithy
           Smithy::Client::Plugins::HttpApiKeyAuth => { require_path: 'smithy-client/plugins/http_api_key_auth' }
         }
       end
+
+      def add_auth_schemes
+        {
+          'smithy.api#httpApiKeyAuth' => {
+            auth_scheme_config_option: :http_api_key_auth_scheme,
+            identity_provider_config_option: :http_api_key_provider,
+            identity_type: Smithy::Client::Identities::HttpApiKey
+          }
+        }
+      end
     end
   end
 end
