+++ gems/smithy/lib/smithy/welds/auth/http_basic_auth.rb
@@ -19,6 +19,16 @@ module Smithy
           Smithy::Client::Plugins::HttpBasicAuth => { require_path: 'smithy-client/plugins/http_basic_auth' }
         }
       end
+
+      def add_auth_schemes
+        {
+          'smithy.api#httpBasicAuth' => {
+            auth_scheme_config_option: :http_basic_auth_scheme,
+            identity_provider_config_option: :http_login_provider,
+            identity_type: Smithy::Client::Identities::HttpLogin
+          }
+        }
+      end
     end
   end
 end
